Move bsdiff4 patching to `Titanic.Helpers`
