+++ Titanic.Updater/PatchUpdateApplier.cs
@@ -1,5 +1,6 @@
 using ICSharpCode.SharpZipLib.Core;
 using ICSharpCode.SharpZipLib.Zip;
+using Titanic.Helpers.Patching;
 
 namespace Titanic.Updater;
